All basic checkers and schedulers have been completed; the output leaves much to be desired, though...

+++ Final Project/src/basics/Task.java
@@ -6,41 +6,41 @@ public class Task
 	 * The total amount (in units of time) that a Task has to run for each
 	 * period.
 	 */
-	int _computation_time = -1;
+	long _computation_time = -1;
 	
 	/**
 	 * The amount (in remaining units of time) that this Task needs to run for
 	 * during this period.
 	 */
-	int _remaining_comp_time = -1;
+	long _remaining_comp_time = -1;
 	
 	/**
 	 * The length (in units of time) of this Task's period.
 	 */
-	int _period = -1;
+	long _period = -1;
 	
 	/**
 	 * The deadline (relative to the start of each period) that this Task needs
 	 * to meet.
 	 */
-	int _deadline = -1;
+	long _deadline = -1;
 	
 	/**
 	 * The deadline (relative to the start of the current period) that this Task
 	 * needs to meet.
 	 */
-	int _next_deadline = -1;
+	long _next_deadline = -1;
 	
 	/**
 	 * The instant (represented in units of time) that the next period of this
 	 * Task begins.
 	 */
-	int _next_period = -1;
+	long _next_period = -1;
 	
 	/**
 	 * the relative value (used in HVDT scheduling) of this Task.
 	 */
-	int _value = -1;
+	double _value = -1;
 	
 	/**
 	 * The name of the Task.
@@ -57,35 +57,35 @@ public class Task
 	 * @param period
 	 *            The length of the period of the Task.
 	 */
-	public Task(String name, int computation_time, int period)
+	public Task(String name, long computation_time, long period)
 	{
 		_name = name;
 		_value = _remaining_comp_time = _computation_time = computation_time;
 		_next_deadline = _deadline = _next_period = _period = period;
 	}
 	
-	public Task(String name, int computation_time, int period, int deadline)
+	public Task(String name, long computation_time, long period, long deadline)
 	{
 		this(name, computation_time, period);
 		_next_deadline = _deadline = deadline;
 	}
 	
-//	@Override
-//	public Task clone()
-//	{
-//		Task ret = new Task(_name, _computation_time, _period);
-//		ret._deadline = _deadline;
-//		ret._next_deadline = _next_deadline;
-//		ret.setValue(_value);
-//		return ret;
-//	}
+	// @Override
+	// public Task clone()
+	// {
+	// Task ret = new Task(_name, _computation_time, _period);
+	// ret._deadline = _deadline;
+	// ret._next_deadline = _next_deadline;
+	// ret.setValue(_value);
+	// return ret;
+	// }
 	
 	/**
 	 * Get the computational need of the Task.
 	 * 
 	 * @return The computational need of the Task.
 	 */
-	public int getComputationTime()
+	public long getComputationTime()
 	{
 		return _computation_time;
 	}
@@ -95,7 +95,7 @@ public class Task
 	 * 
 	 * @return The deadline of the Task.
 	 */
-	public int getDeadline()
+	public long getDeadline()
 	{
 		return _deadline;
 	}
@@ -107,7 +107,7 @@ public class Task
 	 *            The current unit of time.
 	 * @return The laxity of the Task.
 	 */
-	public int getLaxity(int time)
+	public long getLaxity(long time)
 	{
 		return _next_deadline - (time + _remaining_comp_time);
 	}
@@ -127,7 +127,7 @@ public class Task
 	 * 
 	 * @return The current deadline of the Task.
 	 */
-	public int getNextDeadline()
+	public long getNextDeadline()
 	{
 		return _next_deadline;
 	}
@@ -137,7 +137,7 @@ public class Task
 	 * 
 	 * @return The period of the Task.
 	 */
-	public int getPeriod()
+	public long getPeriod()
 	{
 		return _period;
 	}
@@ -147,7 +147,7 @@ public class Task
 	 * 
 	 * @return The remaining computaional need of the Task.
 	 */
-	public int getRemainingCompTime()
+	public long getRemainingCompTime()
 	{
 		return _remaining_comp_time;
 	}
@@ -157,7 +157,7 @@ public class Task
 	 * 
 	 * @return The relative value of the Task.
 	 */
-	public int getValue()
+	public double getValue()
 	{
 		return _value;
 	}
@@ -167,9 +167,9 @@ public class Task
 	 * 
 	 * @return The value density of the Task.
 	 */
-	public int getValueDensity()
+	public double getValueDensity()
 	{
-		return _value / _computation_time;
+		return ((double) _value) / ((double) _computation_time);
 	}
 	
 	/**
@@ -214,7 +214,7 @@ public class Task
 	 *            The new value of the Task.
 	 * @return The new value of the Task.
 	 */
-	public int setValue(int value)
+	public double setValue(double value)
 	{
 		return _value = value;
 	}
@@ -227,7 +227,7 @@ public class Task
 	 * @param details
 	 * @return false if the Task misses it's deadline, true otherwise.
 	 */
-	public boolean update(int time)
+	public boolean update(long time)
 	{
 		if(_next_deadline <= time)
 		{
