+++ Final Project/src/schedulers/DMS_Scheduler.java
@@ -7,19 +7,19 @@ import basics.Task;
 public class DMS_Scheduler extends Scheduler
 {
 	@Override
-	public int getRelevantValue(Task task, int time)
+	public double getRelevantValue(Task task, long time)
 	{
 		return task.getDeadline();
 	}
 	
 	@Override
-	public String printTaskStatus(Task task, int time)
+	public String printTaskStatus(Task task, long time)
 	{
-		return "Task " + task.getName() + " has a relative deadline of " + getRelevantValue(task, time) + ".";
+		return "Task " + task.getName() + " has a relative deadline of " + (long) getRelevantValue(task, time) + ".";
 	}
 	
 	@Override
-	public List<Task> prune(List<Task> list, int time)
+	public List<Task> prune(List<Task> list, long time)
 	{
 		return lowestSort(list, time);
 	}
