+++ Final Project/src/schedulers/EDF_Scheduler.java
@@ -8,26 +8,26 @@ import basics.Task;
 public class EDF_Scheduler extends Scheduler
 {
 	@Override
-	public int getRelevantValue(Task task, int time)
+	public double getRelevantValue(Task task, long time)
 	{
 		return task.getNextDeadline();
 	}
 	
 	@Override
-	public String printTaskStatus(Task task, int time)
+	public String printTaskStatus(Task task, long time)
 	{
-		return "Task " + task.getName() + " has a deadline at " + getRelevantValue(task, time) + ".";
+		return "Task " + task.getName() + " has a deadline at " + (long) getRelevantValue(task, time) + ".";
 	}
 	
 	@Override
-	public List<Task> prune(List<Task> list, int time)
+	public List<Task> prune(List<Task> list, long time)
 	{
 		List<Task> ans = lowestSort(list, time);
 		
 		if(ans.size() < 2)
 			return ans;
 		
-		int laxity = ans.get(0).getLaxity(time);
+		double laxity = ans.get(0).getLaxity(time);
 		
 		for(Task task : ans)
 			laxity = Math.min(laxity, task.getLaxity(time));
