+++ Final Project/src/schedulers/HVDF_Scheduler.java
@@ -6,21 +6,21 @@ import basics.Task;
 
 public class HVDF_Scheduler extends Scheduler
 {
-    @Override
-    public int getRelevantValue(Task task, int time)
-    {
-	return task.getValueDensity();
-    }
-
-    @Override
-    public String printTaskStatus(Task task, int time)
-    {
-	return "Task " + task.getName() + " has a value density of " + getRelevantValue(task, time) + ".";
-    }
-
-    @Override
-    public List<Task> prune(List<Task> list, int time)
-    {
-	return highestSort(list, time);
-    }
+	@Override
+	public double getRelevantValue(Task task, long time)
+	{
+		return task.getValueDensity();
+	}
+	
+	@Override
+	public String printTaskStatus(Task task, long time)
+	{
+		return "Task " + task.getName() + " has a value density of " + getRelevantValue(task, time) + ".";
+	}
+	
+	@Override
+	public List<Task> prune(List<Task> list, long time)
+	{
+		return highestSort(list, time);
+	}
 }
