+++ Final Project/src/schedulers/LLF_Scheduler.java
@@ -8,26 +8,26 @@ import basics.Task;
 public class LLF_Scheduler extends Scheduler
 {
 	@Override
-	public int getRelevantValue(Task task, int time)
+	public double getRelevantValue(Task task, long time)
 	{
 		return task.getLaxity(time);
 	}
 	
 	@Override
-	public String printTaskStatus(Task task, int time)
+	public String printTaskStatus(Task task, long time)
 	{
-		return "Task " + task.getName() + " has a laxity of " + getRelevantValue(task, time) + ".";
+		return "Task " + task.getName() + " has a laxity of " + (long) getRelevantValue(task, time) + ".";
 	}
 	
 	@Override
-	public List<Task> prune(List<Task> list, int time)
+	public List<Task> prune(List<Task> list, long time)
 	{
 		List<Task> ans = lowestSort(list, time);
 		
 		if(ans.size() < 2)
 			return ans;
 		
-		int deadline = ans.get(0).getNextDeadline();
+		long deadline = ans.get(0).getNextDeadline();
 		
 		for(Task task : ans)
 			deadline = Math.min(deadline, task.getNextDeadline());
